requirements.txt

Flask==3.1.2
pymongo==4.15.2
selenium==4.35.0
webdriver-manager==4.0.2
python-dotenv==1.1.1
google-search-results==2.4.2
openai==1.108.1
beautifulsoup4==4.13.5
requests>=2.31.0

